+++ _2DCraft/src/Items/Sapling.java
@@ -2,7 +2,7 @@ package Items;
 
 public class Sapling extends Item{
 
-    public Sapling(double id, String name) {
+    public Sapling(int id, String name) {
         super(id, name);
     }
     @Override
